refactor(igx-grid): Fix test import paths
