+++ projects/igniteui-angular/src/lib/grids/tree-grid/tree-grid-integration.spec.ts
@@ -19,7 +19,7 @@ import { IgxToggleModule } from '../../directives/toggle/toggle.directive';
 import { IgxNumberFilteringOperand } from '../../data-operations/filtering-condition';
 import { IgxHierarchicalTransactionService } from '../../services/transaction/igx-hierarchical-transaction';
 import { IgxGridTransaction } from '../grid-base.component';
-import { IgxGridCellComponent, GridSelectionMode } from '../grid';
+import { IgxGridCellComponent } from '../grid';
 
 const CSS_CLASS_BANNER = 'igx-banner';
 const CSS_CLASS_ROW_EDITED = 'igx-grid__tr--edited';
